// src/generated/models/Cr720_wfrdpricelistsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr720_wfrdpricelistsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr720_wfrdpricelistsstatecode = keyof typeof Cr720_wfrdpricelistsstatecode;
export const Cr720_wfrdpricelistsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr720_wfrdpricelistsstatuscode = keyof typeof Cr720_wfrdpricelistsstatuscode;

export interface Cr720_wfrdpricelistsBase {
  cr720_classification?: string;
  cr720_contractnumber?: string;
  cr720_country?: string;
  cr720_customer?: string;
  cr720_description?: string;
  cr720_gbukey?: string;
  cr720_id?: string;
  cr720_item?: number;
  cr720_latestpossibledeliverydate?: number;
  cr720_ld_aro_fca_location?: string;
  cr720_leadtimeday?: number;
  cr720_plname?: string;
  cr720_qty?: number;
  cr720_unitpriceusd?: number;
  cr720_version?: string;
  cr720_wfrdpn?: string;
  cr720_wfrdpricelistid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr720_wfrdpricelistsstatecode;
  statuscode?: Cr720_wfrdpricelistsstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Cr720_wfrdpricelists extends Cr720_wfrdpricelistsBase {
  cr720_attachments?: string;
  cr720_attachments_name?: string;
  cr720_unitpriceusd_base?: number;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}
export type Cr720_wfrdpricelistsFileColumnName = 'cr720_attachments';

export type Cr720_wfrdpricelistsUploadColumnName = Cr720_wfrdpricelistsFileColumnName;